server/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
anyio==4.6.0
APScheduler==3.10.4
attrs==24.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
certstream==1.12
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
feedparser==6.0.11
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2024.9.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.1.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pluginbase==1.0.1
protobuf==5.28.2
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pycurl==7.45.3
Pygments==2.18.0
pyparsing==3.2.0b3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
python-whois==0.9.4
pytz==2024.1
PyYAML==6.0.2
redis==5.0.5
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
safetensors==0.4.5
selenium==4.22.0
setuptools==70.3.0
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
Telethon==1.36.0
termcolor==2.4.0
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
wafw00f==2.2.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wfuzz==3.1.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
